new scripts directory and OIMS2fits creates a single file per oims file
